Wrap tactics in ElabM in OpResolver
The outer ElabM handles lookup, while the tactic (which contains an
ElabM) handles generating the core terms.

Note that this will get complicated when it comes to declarations
because name resolution depends on resolving the type of the expression
in an `open` declaration. So this change may have to be reverted.